feat(zoom): remove webinar and update list tool requirement
