+++ mcp_servers/zoom/requirements.txt
@@ -5,4 +5,3 @@ uvicorn[standard]
 click
 typing-extensions
 httpx
-zoom-sdk-python
