+++ mcp_servers/zoom/server.py
@@ -15,266 +15,51 @@ from starlette.responses import Response
 from starlette.routing import Mount, Route
 from starlette.types import Receive, Scope, Send
 
+import uvicorn
+
 from tools import (
     zoom_access_token_context,
-    zoom_create_meeting,
-    zoom_get_meeting,
-    zoom_update_meeting,
-    zoom_delete_meeting,
-    zoom_list_meetings,
-    zoom_get_meeting_participants,
-    zoom_create_webinar,
-    zoom_get_user,
-    zoom_list_users,
     validate_access_token,
 )
+
+from list import get_tool_list
+from call import call_tool
 
 logger = logging.getLogger("zoom-mcp-server")
 logging.basicConfig(level=logging.INFO)
 
-ZOOM_MCP_SERVER_PORT = int(os.getenv("ZOOM_MCP_SERVER_PORT", "5000"))
+
+
+class ZoomMCPServer:
+    """Zoom MCP Server with organized tool management."""
+    
+    def __init__(self):
+        self.server = Server("zoom-mcp-server")
+        self._register_handlers()
+    
+    def _register_handlers(self):
+        """Register the tool list and call handlers with the MCP server."""
+        
+        @self.server.list_tools()
+        async def list_tools() -> list[types.Tool]:
+            return get_tool_list()
+
+        @self.server.call_tool()
+        async def call_tool_handler(name: str, arguments: Dict[str, Any]) -> list[types.TextContent]:
+            return await call_tool(name, arguments)
 
 
 @click.command()
-@click.option("--port", default=ZOOM_MCP_SERVER_PORT, help="Port to listen on for HTTP")
+@click.option("--port", default=5000, help="Port to listen on for HTTP")
 @click.option("--log-level", default="INFO", help="Logging level (DEBUG, INFO, WARNING, ERROR, CRITICAL)")
 @click.option("--json-response", is_flag=True, default=True, help="Enable JSON responses for StreamableHTTP")
 def main(port: int, log_level: str, json_response: bool) -> int:
     """Zoom MCP server with SSE + StreamableHTTP transports using OAuth access tokens."""
     logging.getLogger().setLevel(getattr(logging, log_level.upper(), logging.INFO))
-    app = Server("zoom-mcp-server")
-
-    # ----------------------------- Tool Registry -----------------------------#
-    @app.list_tools()
-    async def list_tools() -> list[types.Tool]:
-        return [
-            types.Tool(
-                name="zoom_create_meeting",
-                description=(
-                    "Create a new Zoom meeting with customizable settings like topic, time, duration, "
-                    "and video/audio preferences.\n"
-                    "Prompt hints: 'schedule a meeting', 'create zoom call', 'set up team standup'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "required": ["topic"],
-                    "properties": {
-                        "user_id": {"type": "string", "description": "User ID or 'me' for current user", "default": "me"},
-                        "topic": {"type": "string", "description": "Meeting topic/title"},
-                        "type": {"type": "integer", "description": "Meeting type: 1=instant, 2=scheduled, 3=recurring", "default": 2},
-                        "start_time": {"type": "string", "description": "Start time in ISO format (YYYY-MM-DDTHH:MM:SSZ)"},
-                        "duration": {"type": "integer", "description": "Meeting duration in minutes", "default": 60},
-                        "timezone": {"type": "string", "description": "Timezone (e.g., 'America/New_York')", "default": "UTC"},
-                        "password": {"type": "string", "description": "Meeting password"},
-                        "host_video": {"type": "boolean", "description": "Host video on/off", "default": True},
-                        "participant_video": {"type": "boolean", "description": "Participant video on/off", "default": True},
-                        "join_before_host": {"type": "boolean", "description": "Allow joining before host", "default": False},
-                        "mute_upon_entry": {"type": "boolean", "description": "Mute participants on entry", "default": False},
-                        "watermark": {"type": "boolean", "description": "Add watermark", "default": False},
-                        "use_pmi": {"type": "boolean", "description": "Use Personal Meeting ID", "default": False},
-                        "approval_type": {"type": "integer", "description": "Approval type: 0=automatically, 1=manually, 2=no registration", "default": 0},
-                        "audio": {"type": "string", "description": "Audio options: 'both', 'telephony', 'computer_audio'", "default": "both"},
-                        "auto_recording": {"type": "string", "description": "Auto recording: 'none', 'local', 'cloud'", "default": "none"}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_get_meeting",
-                description=(
-                    "Retrieve detailed information about a specific Zoom meeting including settings, "
-                    "participants, and meeting status.\n"
-                    "Prompt hints: 'get meeting details', 'show meeting info', 'check meeting settings'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "required": ["meeting_id"],
-                    "properties": {
-                        "meeting_id": {"type": "string", "description": "Zoom meeting ID"}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_update_meeting",
-                description=(
-                    "Update meeting details like topic, time, duration, and settings for an existing Zoom meeting.\n"
-                    "Prompt hints: 'change meeting time', 'update meeting topic', 'modify meeting settings'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "required": ["meeting_id"],
-                    "properties": {
-                        "meeting_id": {"type": "string", "description": "Zoom meeting ID"},
-                        "topic": {"type": "string", "description": "New meeting topic/title"},
-                        "type": {"type": "integer", "description": "Meeting type: 1=instant, 2=scheduled, 3=recurring"},
-                        "start_time": {"type": "string", "description": "New start time in ISO format"},
-                        "duration": {"type": "integer", "description": "New meeting duration in minutes"},
-                        "timezone": {"type": "string", "description": "New timezone"},
-                        "password": {"type": "string", "description": "New meeting password"},
-                        "settings": {"type": "object", "description": "Meeting settings to update"}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_delete_meeting",
-                description=(
-                    "Delete a Zoom meeting permanently. This action cannot be undone.\n"
-                    "Prompt hints: 'cancel meeting', 'delete zoom call', 'remove meeting'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "required": ["meeting_id"],
-                    "properties": {
-                        "meeting_id": {"type": "string", "description": "Zoom meeting ID to delete"}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_list_meetings",
-                description=(
-                    "List all meetings for a specific user with pagination support and filtering options.\n"
-                    "Prompt hints: 'show my meetings', 'list all meetings', 'get meeting history'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "properties": {
-                        "user_id": {"type": "string", "description": "User ID or 'me' for current user", "default": "me"},
-                        "type": {"type": "string", "description": "Meeting type filter: 'scheduled', 'live', 'upcoming'"},
-                        "page_size": {"type": "integer", "description": "Number of meetings per page", "default": 30},
-                        "page_number": {"type": "integer", "description": "Page number for pagination", "default": 1},
-                        "next_page_token": {"type": "string", "description": "Token for next page"}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_get_meeting_participants",
-                description=(
-                    "Get the list of participants for a specific meeting including join/leave times and duration.\n"
-                    "Prompt hints: 'who joined the meeting', 'get participant list', 'show meeting attendees'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "required": ["meeting_id"],
-                    "properties": {
-                        "meeting_id": {"type": "string", "description": "Zoom meeting ID"}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_create_webinar",
-                description=(
-                    "Create a new Zoom webinar with advanced settings for large audience events.\n"
-                    "Prompt hints: 'create webinar', 'schedule online event', 'set up large meeting'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "required": ["topic"],
-                    "properties": {
-                        "user_id": {"type": "string", "description": "User ID or 'me' for current user", "default": "me"},
-                        "topic": {"type": "string", "description": "Webinar topic/title"},
-                        "type": {"type": "integer", "description": "Webinar type: 5=scheduled, 6=recurring, 9=recurring with fixed time", "default": 5},
-                        "start_time": {"type": "string", "description": "Start time in ISO format"},
-                        "duration": {"type": "integer", "description": "Webinar duration in minutes", "default": 60},
-                        "timezone": {"type": "string", "description": "Timezone", "default": "UTC"},
-                        "password": {"type": "string", "description": "Webinar password"},
-                        "host_video": {"type": "boolean", "description": "Host video on/off", "default": True},
-                        "panelists_video": {"type": "boolean", "description": "Panelists video on/off", "default": True},
-                        "practice_session": {"type": "boolean", "description": "Enable practice session", "default": False},
-                        "hd_video": {"type": "boolean", "description": "HD video quality", "default": True},
-                        "audio": {"type": "string", "description": "Audio options", "default": "both"},
-                        "auto_recording": {"type": "string", "description": "Auto recording", "default": "none"},
-                        "alternative_hosts": {"type": "string", "description": "Alternative host email addresses"},
-                        "close_registration": {"type": "boolean", "description": "Close registration", "default": False},
-                        "show_share_button": {"type": "boolean", "description": "Show share button", "default": True},
-                        "allow_multiple_devices": {"type": "boolean", "description": "Allow multiple devices", "default": False},
-                        "on_demand": {"type": "boolean", "description": "On-demand webinar", "default": False},
-                        "global_dial_in_countries": {"type": "array", "description": "Countries for dial-in"},
-                        "contact_name": {"type": "string", "description": "Contact person name"},
-                        "contact_email": {"type": "string", "description": "Contact person email"},
-                        "registrants_restrict_number": {"type": "boolean", "description": "Restrict registrants", "default": False},
-                        "registrants_email_notification": {"type": "boolean", "description": "Email notifications", "default": True}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_get_user",
-                description=(
-                    "Get detailed information about a specific Zoom user including account type, "
-                    "settings, and permissions.\n"
-                    "Prompt hints: 'get user info', 'show user details', 'check user account'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "properties": {
-                        "user_id": {"type": "string", "description": "User ID or 'me' for current user", "default": "me"}
-                    }
-                }
-            ),
-            types.Tool(
-                name="zoom_list_users",
-                description=(
-                    "List all users in the Zoom account with pagination and filtering options.\n"
-                    "Prompt hints: 'show all users', 'list team members', 'get user directory'"
-                ),
-                inputSchema={
-                    "type": "object",
-                    "properties": {
-                        "status": {"type": "string", "description": "User status filter: 'active', 'inactive', 'pending'"},
-                        "role_id": {"type": "string", "description": "Role ID filter"},
-                        "page_size": {"type": "integer", "description": "Number of users per page", "default": 30},
-                        "page_number": {"type": "integer", "description": "Page number for pagination", "default": 1},
-                        "next_page_token": {"type": "string", "description": "Token for next page"}
-                    }
-                }
-            ),
-        ]
-
-    # ---------------------------- Tool Dispatcher ----------------------------#
-    @app.call_tool()
-    async def call_tool(name: str, arguments: Dict[str, Any]) -> list[types.TextContent]:
-        logger.info(f"call_tool: {name}")
-        logger.debug(f"raw arguments: {json.dumps(arguments, indent=2)}")
-
-        try:
-            if name == "zoom_create_meeting":
-                result = await zoom_create_meeting(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_get_meeting":
-                result = await zoom_get_meeting(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_update_meeting":
-                result = await zoom_update_meeting(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_delete_meeting":
-                result = await zoom_delete_meeting(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_list_meetings":
-                result = await zoom_list_meetings(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_get_meeting_participants":
-                result = await zoom_get_meeting_participants(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_create_webinar":
-                result = await zoom_create_webinar(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_get_user":
-                result = await zoom_get_user(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            if name == "zoom_list_users":
-                result = await zoom_list_users(arguments)
-                return [types.TextContent(type="text", text=json.dumps(result, indent=2))]
-
-            return [types.TextContent(type="text", text=f"Unknown tool: {name}")]
-        except Exception as e:
-            logger.exception(f"Error executing tool {name}: {e}")
-            return [types.TextContent(type="text", text=f"Error: {str(e)}")]
+    
+    # Initialize the Zoom MCP server
+    zoom_server = ZoomMCPServer()
+    app = zoom_server.server
 
     # ------------------------------- Transports ------------------------------#
     sse = SseServerTransport("/messages/")
@@ -406,11 +191,11 @@ def main(port: int, log_level: str, json_response: bool) -> int:
     logger.info(f"  - StreamableHTTP endpoint: http://localhost:{port}/mcp")
     logger.info("  - OAuth authentication required for all endpoints")
 
-    import uvicorn
     uvicorn.run(starlette_app, host="0.0.0.0", port=port)
 
     return 0
 
 
 if __name__ == "__main__":
+
     exit(main())
