+++ mcp_servers/zoom/test_server.py
@@ -10,7 +10,7 @@ import httpx
 from typing import Dict, Any
 
 # Test configuration - Replace with your actual Zoom OAuth access token
-TEST_ACCESS_TOKEN = "your_zoom_oauth_access_token_here"
+TEST_ACCESS_TOKEN = "eyJzdiI6IjAwMDAwMiIsImFsZyI6IkhTNTEyIiwidiI6IjIuMCIsImtpZCI6ImJkZjk3N2M3LWQwNGEtNDhlNi04OTIxLTg3Y2FlODBhMmY0OSJ9.eyJhdWQiOiJodHRwczovL29hdXRoLnpvb20udXMiLCJ1aWQiOiJJbURhbDBpX1E2cVpzcnhvMmdITi1BIiwidmVyIjoxMCwiYXVpZCI6ImYzODMwMDIzZGEyYmU5MWUxOGY0OGM0MWZiOTkwYTBmZTkxMDAyMWQ5ZmNjNmI0YTc4M2UyNzU5MTZkNWI3MDQiLCJuYmYiOjE3NTU4NDkxMjIsImNvZGUiOiJBd2J6MjNEMVQyV0pUbjlqbUtDYy13Mll0ZmlWc0pqU0YiLCJpc3MiOiJ6bTpjaWQ6Q0VMS2xXYlNUZmVPejF4TzAwcFBnIiwiZ25vIjowLCJleHAiOjE3NTU4NTI3MjIsInR5cGUiOjMsImlhdCI6MTc1NTg0OTEyMiwiYWlkIjoiNHhXX1RjTkxTMWVhOXZERkpIdTNSQSJ9.5ioyydfWUIL9OQFiM5bKMThgoO32O1u_GK8jSe3va-Km8X1nwPGeD28MA6wy04YQwPB3HLVpbvYuHwdt4jqfpA"
 
 async def test_zoom_api_direct():
     """Test Zoom API directly to verify OAuth access token."""
