+++ mcp_servers/zoom/tools/__init__.py
@@ -1,4 +1,4 @@
-from .auth import zoom_api_key_context, get_zoom_client
+from .auth import zoom_access_token_context, get_zoom_client, validate_access_token
 from .meetings import (
     zoom_create_meeting,
     zoom_get_meeting,
@@ -7,13 +7,13 @@ from .meetings import (
     zoom_list_meetings,
     zoom_get_meeting_participants,
 )
-from .webinars import zoom_create_webinar
 from .users import zoom_get_user, zoom_list_users
 
 __all__ = [
     # Auth/context
-    "zoom_api_key_context",
+    "zoom_access_token_context",
     "get_zoom_client",
+    "validate_access_token",
 
     # Meeting tools
     "zoom_create_meeting",
@@ -22,10 +22,7 @@ __all__ = [
     "zoom_delete_meeting",
     "zoom_list_meetings",
     "zoom_get_meeting_participants",
-
-    # Webinar tools
-    "zoom_create_webinar",
-
+    
     # User tools
     "zoom_get_user",
     "zoom_list_users",
